refactor(parser): slim expr/mod.rs by extracting submodules (§7-8)
Reduce the 2885-line parser/expr/mod.rs (over the 500-line limit) to a thin
343-line root by moving the WhateverCode helper functions and the large
test module into new sibling submodules, all under 500 lines. mod.rs keeps
its existing mod/use/re-export declarations, the memo statics, and the core
expression entry points; new submodules are re-exported preserving original
visibility, so no external caller changes.

- mod.rs (343)                - root: entry points, memo, mod/use/re-exports
- whatever.rs (392)           - WhateverCode detection/inspection
- whatever_replace.rs (353)   - placeholder replacement
- whatever_wrap.rs (293)      - WhateverCode wrapping/composition
- tests.rs / tests_errors.rs / tests_meta.rs / tests_postfix.rs (445/364/355/381)
                              - the former inline #[cfg(test)] mod, split by theme

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; cross-submodule helpers bumped to pub(crate).
The fn multiset is byte-identical. No behavior change.
