+++ src/parser/expr/mod.rs
@@ -2,6 +2,18 @@ mod operators;
 mod postfix;
 pub(crate) mod precedence;
 mod precedence_meta_ops;
+mod whatever;
+mod whatever_replace;
+mod whatever_wrap;
+
+#[cfg(test)]
+mod tests;
+#[cfg(test)]
+mod tests_errors;
+#[cfg(test)]
+mod tests_meta;
+#[cfg(test)]
+mod tests_postfix;
 
 use super::memo::{MemoEntry, MemoStats, ParseMemo};
 use super::parse_result::PResult;
@@ -16,6 +28,18 @@ use super::helpers::ws;
 pub(in crate::parser) use postfix::postfix_expr_continue;
 pub(in crate::parser) use postfix::{QuotedMethodName, parse_quoted_method_name};
 use precedence::ternary;
+
+// Re-exports for WhateverCode detection (`whatever.rs`).
+pub(in crate::parser) use whatever::should_wrap_whatevercode;
+pub(super) use whatever::{contains_whatever, is_whatever};
+pub(crate) use whatever::{count_whatever, expr_contains_topic, exprs_structurally_eq};
+
+// Re-exports for WhateverCode wrapping (`whatever_wrap.rs`).
+pub(in crate::parser) use whatever_wrap::wrap_whatevercode;
+use whatever_wrap::{try_wrap_whatevercode_call_chain, wrap_composition_operands};
+
+// Re-exports for WhateverCode body construction (`whatever_replace.rs`).
+pub(crate) use whatever_replace::{replace_whatever_numbered, replace_whatever_single};
 
 thread_local! {
     static EXPR_MEMO_TLS: RefCell<HashMap<(usize, usize), MemoEntry<Expr>>> = RefCell::new(HashMap::new());
@@ -316,2570 +340,4 @@ pub(in crate::parser) fn parse_fat_arrow_value(input: &str) -> PResult<'_, Expr>
 
 pub(in crate::parser) fn term_expr(input: &str) -> PResult<'_, Expr> {
     postfix::prefix_expr(input)
-}
-
-pub(in crate::parser) fn should_wrap_whatevercode(expr: &Expr) -> bool {
-    if !contains_whatever(expr) || is_whatever(expr) {
-        return false;
-    }
-    if contains_xx_with_bare_whatever(expr) {
-        return false;
-    }
-    match expr {
-        // SmartMatch: Whatever on RHS is handled at runtime (autoprime).
-        // LHS compound Whatever curries (e.g. `*.abs ~~ Code`), but bare
-        // `* ~~ Type` currying is handled in the precedence parser where we can
-        // distinguish a true bare * operand from a parenthesized ((*)).
-        Expr::Binary {
-            op: TokenKind::SmartMatch,
-            ..
-        } => false,
-        // BangTilde (!~~): bare * on LHS SHOULD curry (e.g. `* !~~ Int`), because
-        // `!~~` is not Whatever-aware in Raku.
-        Expr::Binary {
-            op: TokenKind::BangTilde,
-            left,
-            ..
-        } => contains_whatever(left),
-        Expr::Binary {
-            op: TokenKind::Ident(name),
-            ..
-        } if name == "o" => false,
-        // List replication `xx` does not Whatever-curry a *bare* `*` operand: a
-        // standalone `*` is the Whatever value, repeated literally. `* xx 2` is
-        // `(*, *)`; `1 xx *`/`1 x *` is the infinite-repeat form. None wrap into a
-        // WhateverCode. (A compound operand like `(*+1) xx 2` is already wrapped at
-        // the parenthesis, so its `left` is a Lambda, not a bare Whatever. Note
-        // string replication `* x 2` DOES curry into a WhateverCode in Raku, so
-        // only the right-Whatever form is exempt for `x`.)
-        Expr::Binary {
-            op: TokenKind::Ident(name),
-            left,
-            right,
-        } if ((name == "x" || name == "xx") && matches!(&**right, Expr::Whatever))
-            || (name == "xx" && matches!(&**left, Expr::Whatever)) =>
-        {
-            false
-        }
-        _ => true,
-    }
-}
-
-fn contains_xx_with_bare_whatever(expr: &Expr) -> bool {
-    match expr {
-        Expr::Binary {
-            left,
-            op: TokenKind::Ident(name),
-            right,
-            ..
-        } => {
-            // `xx` with a *bare* `*` on either side is a literal repetition of the
-            // Whatever value (`* xx 2` → `(*, *)`, `1 xx *` → infinite), not a curry
-            // point — so an enclosing postfix (`(* xx 3).elems`) must not wrap the
-            // whole chain into a WhateverCode either. (`x`/`xx` with a right `*` was
-            // already exempt; string `* x 2` DOES curry, so only `xx` exempts left.)
-            ((name == "xx" || name == "x") && matches!(&**right, Expr::Whatever))
-                || (name == "xx" && matches!(&**left, Expr::Whatever))
-                || contains_xx_with_bare_whatever(left)
-                || contains_xx_with_bare_whatever(right)
-        }
-        Expr::Unary { expr, .. } => contains_xx_with_bare_whatever(expr),
-        Expr::MethodCall { target, args, .. } => {
-            contains_xx_with_bare_whatever(target)
-                || args.iter().any(contains_xx_with_bare_whatever)
-        }
-        Expr::Index { target, index, .. } => {
-            contains_xx_with_bare_whatever(target) || contains_xx_with_bare_whatever(index)
-        }
-        Expr::Call { args, .. } => args.iter().any(contains_xx_with_bare_whatever),
-        Expr::CallOn { target, args } => {
-            contains_xx_with_bare_whatever(target)
-                || args.iter().any(contains_xx_with_bare_whatever)
-        }
-        Expr::ArrayLiteral(items) | Expr::BracketArray(items, _) => {
-            items.iter().any(contains_xx_with_bare_whatever)
-        }
-        Expr::CaptureLiteral(items) => items.iter().any(contains_xx_with_bare_whatever),
-        _ => false,
-    }
-}
-
-fn wrap_composition_operands(expr: Expr) -> Expr {
-    match expr {
-        Expr::Binary { left, op, right } => {
-            let left = wrap_composition_operands(*left);
-            let right = wrap_composition_operands(*right);
-            if matches!(&op, TokenKind::Ident(name) if name == "o") {
-                let mut bare_count = 0usize;
-                if is_whatever(&left) {
-                    bare_count += 1;
-                }
-                if is_whatever(&right) {
-                    bare_count += 1;
-                }
-                if bare_count > 0 {
-                    let mut params = Vec::new();
-                    let mut param_defs = Vec::new();
-                    let left_expr = if is_whatever(&left) {
-                        let name = format!("__wc_{}", params.len());
-                        params.push(name.clone());
-                        param_defs.push(make_wc_param(name.clone()));
-                        Expr::Var(name)
-                    } else if should_wrap_whatevercode(&left) {
-                        wrap_whatevercode(&left)
-                    } else {
-                        left
-                    };
-                    let right_expr = if is_whatever(&right) {
-                        let name = format!("__wc_{}", params.len());
-                        params.push(name.clone());
-                        param_defs.push(make_wc_param(name.clone()));
-                        Expr::Var(name)
-                    } else if should_wrap_whatevercode(&right) {
-                        wrap_whatevercode(&right)
-                    } else {
-                        right
-                    };
-                    let body_expr = Expr::Binary {
-                        left: Box::new(left_expr),
-                        op,
-                        right: Box::new(right_expr),
-                    };
-                    if params.len() == 1 {
-                        return Expr::Lambda {
-                            param: params[0].clone(),
-                            body: vec![Stmt::Expr(body_expr)],
-                            is_whatever_code: false,
-                        };
-                    }
-                    return Expr::AnonSubParams {
-                        params,
-                        param_defs,
-                        return_type: None,
-                        body: vec![Stmt::Expr(body_expr)],
-                        is_rw: false,
-                        is_whatever_code: false,
-                    };
-                }
-                let left_wrapped = if should_wrap_whatevercode(&left) {
-                    wrap_whatevercode(&left)
-                } else {
-                    left
-                };
-                let right_wrapped = if should_wrap_whatevercode(&right) {
-                    wrap_whatevercode(&right)
-                } else {
-                    right
-                };
-                Expr::Binary {
-                    left: Box::new(left_wrapped),
-                    op,
-                    right: Box::new(right_wrapped),
-                }
-            } else {
-                Expr::Binary {
-                    left: Box::new(left),
-                    op,
-                    right: Box::new(right),
-                }
-            }
-        }
-        Expr::Unary { op, expr } => Expr::Unary {
-            op,
-            expr: Box::new(wrap_composition_operands(*expr)),
-        },
-        Expr::MethodCall {
-            target,
-            name,
-            args,
-            modifier,
-            quoted,
-        } => Expr::MethodCall {
-            target: Box::new(wrap_composition_operands(*target)),
-            name,
-            args: args.into_iter().map(wrap_composition_operands).collect(),
-            modifier,
-            quoted,
-        },
-        Expr::CallOn { target, args } => Expr::CallOn {
-            target: Box::new(wrap_composition_operands(*target)),
-            args: args.into_iter().map(wrap_composition_operands).collect(),
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-            ..
-        } => Expr::Index {
-            target: Box::new(wrap_composition_operands(*target)),
-            index: Box::new(wrap_composition_operands(*index)),
-            is_positional,
-        },
-        other => other,
-    }
-}
-
-fn make_wc_param(name: String) -> crate::ast::ParamDef {
-    crate::ast::ParamDef {
-        name,
-        default: None,
-        multi_invocant: true,
-        required: false,
-        named: false,
-        slurpy: false,
-        double_slurpy: false,
-        onearg: false,
-        sigilless: false,
-        type_constraint: None,
-        literal_value: None,
-        sub_signature: None,
-        where_constraint: None,
-        traits: Vec::new(),
-        optional_marker: false,
-        outer_sub_signature: None,
-        code_signature: None,
-        is_invocant: false,
-        shape_constraints: None,
-    }
-}
-
-/// Try to detect and fix a chain of MethodCalls leading to a CallOn whose target
-/// contains Whatever. If found, wrap only the CallOn target as WhateverCode,
-/// leaving the outer method calls outside the lambda.
-///
-/// Handles patterns like: *.foo().(args).bar().baz()
-/// where only *.foo() should be wrapped as WhateverCode.
-fn try_wrap_whatevercode_call_chain(expr: &Expr) -> Option<Expr> {
-    // Check if this is a MethodCall chain ending at a CallOn with Whatever target
-    match expr {
-        Expr::MethodCall {
-            target,
-            name,
-            args,
-            modifier,
-            quoted,
-        } if !args.iter().any(contains_whatever) => {
-            match target.as_ref() {
-                // Direct: MethodCall -> CallOn -> Whatever-containing target
-                Expr::CallOn {
-                    target: inner_target,
-                    args: call_args,
-                } if should_wrap_whatevercode(inner_target)
-                    && !call_args.iter().any(contains_whatever) =>
-                {
-                    Some(Expr::MethodCall {
-                        target: Box::new(Expr::CallOn {
-                            target: Box::new(wrap_whatevercode(inner_target)),
-                            args: call_args.clone(),
-                        }),
-                        name: *name,
-                        args: args.clone(),
-                        modifier: *modifier,
-                        quoted: *quoted,
-                    })
-                }
-                // Recursive: MethodCall -> MethodCall -> ... -> CallOn
-                inner @ Expr::MethodCall { .. } => {
-                    let wrapped_inner = try_wrap_whatevercode_call_chain(inner)?;
-                    Some(Expr::MethodCall {
-                        target: Box::new(wrapped_inner),
-                        name: *name,
-                        args: args.clone(),
-                        modifier: *modifier,
-                        quoted: *quoted,
-                    })
-                }
-                _ => None,
-            }
-        }
-        _ => None,
-    }
-}
-
-pub(super) fn is_whatever(expr: &Expr) -> bool {
-    matches!(expr, Expr::Whatever)
-}
-
-/// True when `expr` is an already-wrapped WhateverCode closure (produced by a
-/// parenthesized curry such as `(* - 1)`). Such a closure is opaque as a *value*
-/// (e.g. when passed as an argument or stored in a variable), but when it appears
-/// as an *operand* of a further operator/method in the same expression, Raku
-/// composes it into a new, larger WhateverCode (`(* - 1) - 1`, `(^*).roll`,
-/// `1 +< (* - 1)`). The currying machinery (`count_whatever` /
-/// `replace_whatever_*`) already knows how to inline such a closure; this helper
-/// lets the composing operand positions detect it.
-fn is_wrapped_whatevercode(expr: &Expr) -> bool {
-    matches!(
-        expr,
-        Expr::Lambda {
-            is_whatever_code: true,
-            ..
-        } | Expr::AnonSubParams {
-            is_whatever_code: true,
-            ..
-        }
-    )
-}
-
-pub(super) fn contains_whatever(expr: &Expr) -> bool {
-    match expr {
-        e if is_whatever(e) => true,
-        // Don't treat bare * inside range/sequence operators as WhateverCode.
-        // `1..*` is a Range, but `1..*-1` is a WhateverCode.
-        // If an endpoint contains a non-bare Whatever (e.g. `*-1`), the whole
-        // range should become a WhateverCode.
-        Expr::Binary {
-            op:
-                TokenKind::DotDot
-                | TokenKind::DotDotCaret
-                | TokenKind::CaretDotDot
-                | TokenKind::CaretDotDotCaret,
-            left,
-            right,
-        } => {
-            fn endpoint_has_compound_whatever(e: &Expr) -> bool {
-                contains_whatever(e) && !is_whatever(e)
-            }
-            endpoint_has_compound_whatever(left) || endpoint_has_compound_whatever(right)
-        }
-        Expr::Binary {
-            op: TokenKind::DotDotDot | TokenKind::DotDotDotCaret,
-            ..
-        } => false,
-        // SmartMatch/BangTilde: Whatever on the RHS is handled at runtime
-        // (autoprime to WhateverCode). Only check LHS for Whatever.
-        Expr::Binary {
-            op: TokenKind::SmartMatch | TokenKind::BangTilde,
-            left,
-            ..
-        } => contains_whatever(left),
-        // Named FatArrow pairs (colonpairs): `:as(*)` produces `"as" => *` which should
-        // be Pair("as", Whatever), NOT a WhateverCode.  When the left side is a string
-        // literal, skip WhateverCode propagation so the Pair is not auto-curried.
-        Expr::Binary {
-            op: TokenKind::FatArrow,
-            left,
-            ..
-        } if matches!(left.as_ref(), Expr::Literal(Value::Str(_))) => false,
-        // Composition operators `o` and `∘` never auto-curry: their operands are
-        // always treated as callables, so `(* + 1) o (* * 2)` should compose
-        // two WhateverCodes rather than becoming a WhateverCode itself.
-        Expr::Binary {
-            op: TokenKind::Ident(name),
-            ..
-        } if name == "o" || name == "\u{2218}" => false,
-        // `xx` replicates its LHS as a value N times (producing a Seq), so a
-        // WhateverCode LHS is NOT composed — `(* - 1) xx 3` is a Seq of three
-        // WhateverCodes, not a curried WhateverCode. Keep the historical
-        // bare-`*` behavior (only the literal placeholder triggers wrapping).
-        Expr::Binary {
-            op: TokenKind::Ident(name),
-            left,
-            right,
-        } if name == "xx" => contains_whatever(left) || contains_whatever(right),
-        Expr::Binary { left, right, .. } => {
-            contains_whatever(left)
-                || contains_whatever(right)
-                || is_wrapped_whatevercode(left)
-                || is_wrapped_whatevercode(right)
-        }
-        Expr::Unary { expr, .. } | Expr::PostfixOp { expr, .. } => {
-            contains_whatever(expr) || is_wrapped_whatevercode(expr)
-        }
-        // Pseudo-methods (.WHAT, .WHO, .HOW, etc.) are always evaluated immediately
-        // on Whatever, they don't create WhateverCode. e.g. *.WHAT returns (Whatever).
-        Expr::MethodCall { target, name, .. }
-            if matches!(
-                name.resolve().as_str(),
-                "WHAT" | "WHO" | "HOW" | "WHY" | "WHICH" | "WHERE" | "DEFINITE" | "VAR"
-            ) && is_whatever(target) =>
-        {
-            false
-        }
-        Expr::MethodCall { target, .. } | Expr::DynamicMethodCall { target, .. } => {
-            contains_whatever(target) || is_wrapped_whatevercode(target)
-        }
-        Expr::CallOn { target, args } => {
-            // Already-wrapped WhateverCode args (Lambda/AnonSubParams with
-            // is_whatever_code: true) are opaque values, not raw Whatever
-            // placeholders.  Only bare * or compound-Whatever args should
-            // trigger auto-currying of the whole CallOn.
-            contains_whatever(target)
-                || args.iter().any(|a| {
-                    !matches!(
-                        a,
-                        Expr::Lambda {
-                            is_whatever_code: true,
-                            ..
-                        } | Expr::AnonSubParams {
-                            is_whatever_code: true,
-                            ..
-                        }
-                    ) && contains_whatever(a)
-                })
-        }
-        // Only check target, not index: @a[*-1] should NOT make the whole expr a WhateverCode.
-        // The [*-1] subscript handles its own WhateverCode wrapping.
-        Expr::Index { target, .. } => contains_whatever(target) || is_wrapped_whatevercode(target),
-        // User-defined infix operators: `* quack 5`, `3 quack *`, etc.
-        // Exclude flip-flop operators (ff, fff and variants) since `ff *` means
-        // "stay true forever" and `*` should not trigger WhateverCode wrapping.
-        Expr::InfixFunc {
-            name, left, right, ..
-        } => {
-            let is_flipflop = matches!(
-                name.as_str(),
-                "ff" | "fff" | "^ff" | "ff^" | "^ff^" | "^fff" | "fff^" | "^fff^"
-            );
-            if is_flipflop {
-                return false;
-            }
-            contains_whatever(left)
-                || is_wrapped_whatevercode(left)
-                || right
-                    .iter()
-                    .any(|e| contains_whatever(e) || is_wrapped_whatevercode(e))
-        }
-        // R meta-operators with Whatever: `5 R- *` should curry.
-        // X/Z meta-operators with bare * in list contexts mean "extend" rather
-        // than WhateverCode, so only enable for R (reverse) meta-ops.
-        Expr::MetaOp {
-            meta, left, right, ..
-        } if meta == "R" => {
-            contains_whatever(left)
-                || contains_whatever(right)
-                || is_wrapped_whatevercode(left)
-                || is_wrapped_whatevercode(right)
-        }
-        _ => false,
-    }
-}
-
-/// Count the number of distinct Whatever (`*`) placeholders in an expression.
-fn count_whatever(expr: &Expr) -> usize {
-    match expr {
-        e if is_whatever(e) => 1,
-        // A nested, already-wrapped WhateverCode operand (e.g. `(* - 1)` inside
-        // `(* - 1) - 1`) contributes its own placeholder count: a single-param
-        // `_`/`__wc_0` lambda counts as 1, a multi-param one as its arity.
-        Expr::Lambda {
-            is_whatever_code: true,
-            ..
-        } => 1,
-        Expr::AnonSubParams {
-            is_whatever_code: true,
-            params,
-            ..
-        } => params.len(),
-        Expr::Binary {
-            left,
-            op: TokenKind::AndAnd,
-            right,
-        } => {
-            if let (
-                Expr::Binary {
-                    left: ll,
-                    right: lr,
-                    ..
-                },
-                Expr::Binary {
-                    left: rl,
-                    right: rr,
-                    ..
-                },
-            ) = (left.as_ref(), right.as_ref())
-                && exprs_structurally_eq(lr, rl)
-                && count_whatever(lr) > 0
-            {
-                // Chained comparison `a OP m OP b` is expanded to
-                // `(a OP m) && (m OP b)` with the middle `m` duplicated. Count the
-                // shared middle's placeholders once so the WhateverCode arity is
-                // the number of distinct operands, not double the middle.
-                return count_whatever(ll) + count_whatever(lr) + count_whatever(rr);
-            }
-            count_whatever(left) + count_whatever(right)
-        }
-        // For range operators: count Whatever in endpoints only when
-        // the endpoint contains compound Whatever (not bare *).
-        Expr::Binary {
-            op:
-                TokenKind::DotDot
-                | TokenKind::DotDotCaret
-                | TokenKind::CaretDotDot
-                | TokenKind::CaretDotDotCaret,
-            left,
-            right,
-        } => {
-            let lc = if contains_whatever(left) && !is_whatever(left) {
-                count_whatever(left)
-            } else {
-                0
-            };
-            let rc = if contains_whatever(right) && !is_whatever(right) {
-                count_whatever(right)
-            } else {
-                0
-            };
-            lc + rc
-        }
-        Expr::Binary {
-            op: TokenKind::DotDotDot | TokenKind::DotDotDotCaret,
-            ..
-        } => 0,
-        // SmartMatch/BangTilde: Whatever on RHS is handled at runtime; only count LHS.
-        Expr::Binary {
-            op: TokenKind::SmartMatch | TokenKind::BangTilde,
-            left,
-            ..
-        } => count_whatever(left),
-        Expr::Binary { left, right, .. } => count_whatever(left) + count_whatever(right),
-        Expr::Unary { expr, .. } | Expr::PostfixOp { expr, .. } => count_whatever(expr),
-        Expr::MethodCall { target, .. } | Expr::DynamicMethodCall { target, .. } => {
-            count_whatever(target)
-        }
-        Expr::CallOn { target, args } => {
-            count_whatever(target) + args.iter().map(count_whatever).sum::<usize>()
-        }
-        // Only check target, not index (subscript handles its own WhateverCode)
-        Expr::Index { target, .. } => count_whatever(target),
-        // User-defined infix operators
-        Expr::InfixFunc { left, right, .. } => {
-            count_whatever(left) + right.iter().map(count_whatever).sum::<usize>()
-        }
-        // R/X/Z meta-operators. R always curries on a Whatever operand; X/Z
-        // currying is decided at parse time in `container.rs` (a standalone `*`
-        // operand curries, a trailing `*` in a comma-list operand extends), but
-        // once the decision to wrap is made we must count placeholders here so
-        // the WhateverCode gets the right arity.
-        Expr::MetaOp {
-            meta, left, right, ..
-        } if matches!(meta.as_str(), "R" | "X" | "Z") => {
-            count_whatever(left) + count_whatever(right)
-        }
-        _ => 0,
-    }
-}
-
-/// Structural equality of two expressions, used to detect the shared middle
-/// term of an expanded chained comparison (`a OP m OP b` => `(a OP m) && (m OP
-/// b)`). `Expr` cannot derive `PartialEq` (it embeds `Value`), and this only runs
-/// while wrapping a WhateverCode, so a `Debug`-string comparison is sufficient.
-fn exprs_structurally_eq(a: &Expr, b: &Expr) -> bool {
-    format!("{a:?}") == format!("{b:?}")
-}
-
-/// Replace Whatever expressions with numbered parameter variables.
-/// `counter` tracks the next parameter index to assign.
-fn replace_whatever_numbered(expr: &Expr, counter: &mut usize) -> Expr {
-    match expr {
-        e if is_whatever(e) => {
-            let var_name = format!("__wc_{}", counter);
-            *counter += 1;
-            Expr::Var(var_name)
-        }
-        Expr::Binary {
-            left,
-            op: TokenKind::AndAnd,
-            right,
-        } => {
-            if let (
-                Expr::Binary {
-                    left: ll,
-                    op: lop,
-                    right: lr,
-                },
-                Expr::Binary {
-                    left: rl,
-                    op: rop,
-                    right: rr,
-                },
-            ) = (left.as_ref(), right.as_ref())
-                && exprs_structurally_eq(lr, rl)
-                && count_whatever(lr) > 0
-            {
-                // Chained comparison expanded to `(ll OP m) && (m OP rr)`: assign
-                // params left-to-right (ll, then the shared middle once, then rr)
-                // and reuse the same replaced middle in both comparisons so each
-                // distinct operand maps to its own positional argument.
-                let new_ll = replace_whatever_numbered(ll, counter);
-                let new_mid = replace_whatever_numbered(lr, counter);
-                let new_rr = replace_whatever_numbered(rr, counter);
-                return Expr::Binary {
-                    left: Box::new(Expr::Binary {
-                        left: Box::new(new_ll),
-                        op: lop.clone(),
-                        right: Box::new(new_mid.clone()),
-                    }),
-                    op: TokenKind::AndAnd,
-                    right: Box::new(Expr::Binary {
-                        left: Box::new(new_mid),
-                        op: rop.clone(),
-                        right: Box::new(new_rr),
-                    }),
-                };
-            }
-            Expr::Binary {
-                left: Box::new(replace_whatever_numbered(left, counter)),
-                op: TokenKind::AndAnd,
-                right: Box::new(replace_whatever_numbered(right, counter)),
-            }
-        }
-        // Unwrap a nested WhateverCode lambda: reuse its body with renumbered params
-        Expr::Lambda { param, body, .. } if param == "_" => {
-            // Single-param WhateverCode: rename $_ to the next numbered param
-            let var_name = format!("__wc_{}", counter);
-            *counter += 1;
-            if let Some(Stmt::Expr(e)) = body.first() {
-                rename_var(e, "_", &var_name)
-            } else {
-                Expr::Var(var_name)
-            }
-        }
-        Expr::AnonSubParams { params, body, .. }
-            if params.iter().all(|p| p.starts_with("__wc_")) =>
-        {
-            // Multi-param WhateverCode: renumber all params
-            let mut renames = Vec::new();
-            for old_name in params {
-                let new_name = format!("__wc_{}", counter);
-                *counter += 1;
-                renames.push((old_name.clone(), new_name));
-            }
-            let mut body_expr = if let Some(Stmt::Expr(e)) = body.first() {
-                e.clone()
-            } else {
-                return expr.clone();
-            };
-            for (old, new) in &renames {
-                body_expr = rename_var(&body_expr, old, new);
-            }
-            body_expr
-        }
-        // SmartMatch: only replace Whatever on LHS; RHS is handled at runtime.
-        Expr::Binary {
-            left,
-            op: op @ TokenKind::SmartMatch,
-            right,
-        } => Expr::Binary {
-            left: Box::new(replace_whatever_numbered(left, counter)),
-            op: op.clone(),
-            right: right.clone(),
-        },
-        Expr::Binary { left, op, right } => Expr::Binary {
-            left: Box::new(replace_whatever_numbered(left, counter)),
-            op: op.clone(),
-            right: Box::new(replace_whatever_numbered(right, counter)),
-        },
-        Expr::Unary { op, expr } => Expr::Unary {
-            op: op.clone(),
-            expr: Box::new(replace_whatever_numbered(expr, counter)),
-        },
-        Expr::PostfixOp { op, expr } => Expr::PostfixOp {
-            op: op.clone(),
-            expr: Box::new(replace_whatever_numbered(expr, counter)),
-        },
-        Expr::MethodCall {
-            target,
-            name,
-            args,
-            modifier,
-            quoted,
-        } => Expr::MethodCall {
-            target: Box::new(replace_whatever_numbered(target, counter)),
-            name: *name,
-            args: args.clone(),
-            modifier: *modifier,
-            quoted: *quoted,
-        },
-        Expr::DynamicMethodCall {
-            target,
-            name_expr,
-            args,
-            modifier,
-        } => Expr::DynamicMethodCall {
-            target: Box::new(replace_whatever_numbered(target, counter)),
-            name_expr: name_expr.clone(),
-            args: args.clone(),
-            modifier: *modifier,
-        },
-        Expr::CallOn { target, args } => Expr::CallOn {
-            target: Box::new(replace_whatever_numbered(target, counter)),
-            args: args
-                .iter()
-                .map(|a| replace_whatever_numbered(a, counter))
-                .collect(),
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-            ..
-        } => Expr::Index {
-            target: Box::new(replace_whatever_numbered(target, counter)),
-            index: index.clone(),
-            is_positional: *is_positional,
-        },
-        Expr::InfixFunc {
-            name,
-            left,
-            right,
-            modifier,
-        } => Expr::InfixFunc {
-            name: name.clone(),
-            left: Box::new(replace_whatever_numbered(left, counter)),
-            right: right
-                .iter()
-                .map(|a| replace_whatever_numbered(a, counter))
-                .collect(),
-            modifier: modifier.clone(),
-        },
-        Expr::MetaOp {
-            meta,
-            op,
-            left,
-            right,
-        } => Expr::MetaOp {
-            meta: meta.clone(),
-            op: op.clone(),
-            left: Box::new(replace_whatever_numbered(left, counter)),
-            right: Box::new(replace_whatever_numbered(right, counter)),
-        },
-        _ => expr.clone(),
-    }
-}
-
-/// Rename a variable in an expression tree.
-fn rename_var(expr: &Expr, old_name: &str, new_name: &str) -> Expr {
-    match expr {
-        Expr::Var(name) if name == old_name => Expr::Var(new_name.to_string()),
-        Expr::Binary { left, op, right } => Expr::Binary {
-            left: Box::new(rename_var(left, old_name, new_name)),
-            op: op.clone(),
-            right: Box::new(rename_var(right, old_name, new_name)),
-        },
-        Expr::Unary { op, expr } => Expr::Unary {
-            op: op.clone(),
-            expr: Box::new(rename_var(expr, old_name, new_name)),
-        },
-        Expr::PostfixOp { op, expr } => Expr::PostfixOp {
-            op: op.clone(),
-            expr: Box::new(rename_var(expr, old_name, new_name)),
-        },
-        Expr::MethodCall {
-            target,
-            name,
-            args,
-            modifier,
-            quoted,
-        } => Expr::MethodCall {
-            target: Box::new(rename_var(target, old_name, new_name)),
-            name: *name,
-            args: args
-                .iter()
-                .map(|a| rename_var(a, old_name, new_name))
-                .collect(),
-            modifier: *modifier,
-            quoted: *quoted,
-        },
-        Expr::CallOn { target, args } => Expr::CallOn {
-            target: Box::new(rename_var(target, old_name, new_name)),
-            args: args
-                .iter()
-                .map(|a| rename_var(a, old_name, new_name))
-                .collect(),
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-            ..
-        } => Expr::Index {
-            target: Box::new(rename_var(target, old_name, new_name)),
-            index: Box::new(rename_var(index, old_name, new_name)),
-            is_positional: *is_positional,
-        },
-        _ => expr.clone(),
-    }
-}
-
-/// Build a WhateverCode lambda from an expression containing Whatever placeholders.
-pub(in crate::parser) fn wrap_whatevercode(expr: &Expr) -> Expr {
-    if let Expr::CallOn { target, args } = expr
-        && should_wrap_whatevercode(target)
-        && !args.iter().any(contains_whatever)
-    {
-        return Expr::CallOn {
-            target: Box::new(wrap_whatevercode(target)),
-            args: args.clone(),
-        };
-    }
-
-    let wc_count = count_whatever(expr);
-
-    if wc_count <= 1 && !expr_contains_topic(expr) {
-        // Single-arg: use Lambda with param "_" for backward compat
-        // Use a special single-arg replacer that maps * and nested single-arg WC to $_
-        let body_expr = replace_whatever_single(expr);
-        Expr::Lambda {
-            param: "_".to_string(),
-            body: vec![Stmt::Expr(body_expr)],
-            is_whatever_code: true,
-        }
-    } else if wc_count <= 1 {
-        // Single-arg, but expression already contains $_ — use a numbered param
-        // to avoid shadowing the outer $_.
-        let mut counter = 0;
-        let body_expr = replace_whatever_numbered(expr, &mut counter);
-        let param_name = "__wc_0".to_string();
-        Expr::AnonSubParams {
-            params: vec![param_name.clone()],
-            param_defs: vec![crate::ast::ParamDef {
-                name: param_name,
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy: false,
-                sigilless: false,
-                type_constraint: None,
-                literal_value: None,
-                sub_signature: None,
-                where_constraint: None,
-                traits: Vec::new(),
-                double_slurpy: false,
-                onearg: false,
-                optional_marker: false,
-                outer_sub_signature: None,
-                code_signature: None,
-                is_invocant: false,
-                shape_constraints: None,
-            }],
-            return_type: None,
-            body: vec![Stmt::Expr(body_expr)],
-            is_rw: false,
-            is_whatever_code: true,
-        }
-    } else {
-        // Multi-arg: use AnonSubParams with numbered params
-        let mut counter = 0;
-        let body_expr = replace_whatever_numbered(expr, &mut counter);
-        let params: Vec<String> = (0..counter).map(|i| format!("__wc_{}", i)).collect();
-        Expr::AnonSubParams {
-            params: params.clone(),
-            param_defs: params
-                .iter()
-                .map(|name| crate::ast::ParamDef {
-                    name: name.clone(),
-                    default: None,
-                    multi_invocant: true,
-                    required: false,
-                    named: false,
-                    slurpy: false,
-                    sigilless: false,
-                    type_constraint: None,
-                    literal_value: None,
-                    sub_signature: None,
-                    where_constraint: None,
-                    traits: Vec::new(),
-                    double_slurpy: false,
-                    onearg: false,
-                    optional_marker: false,
-                    outer_sub_signature: None,
-                    code_signature: None,
-                    is_invocant: false,
-                    shape_constraints: None,
-                })
-                .collect(),
-            return_type: None,
-            body: vec![Stmt::Expr(body_expr)],
-            is_rw: false,
-            is_whatever_code: true,
-        }
-    }
-}
-
-/// Check if an expression contains a reference to $_ (the topic variable).
-/// Used to determine whether a WhateverCode lambda should avoid using $_ as its param.
-fn expr_contains_topic(expr: &Expr) -> bool {
-    match expr {
-        Expr::Var(name) if name == "_" => true,
-        Expr::Whatever => false,
-        Expr::Binary { left, right, .. } => expr_contains_topic(left) || expr_contains_topic(right),
-        Expr::Unary { expr, .. } | Expr::PostfixOp { expr, .. } => expr_contains_topic(expr),
-        Expr::MethodCall { target, args, .. } => {
-            expr_contains_topic(target) || args.iter().any(expr_contains_topic)
-        }
-        Expr::CallOn { target, args } => {
-            expr_contains_topic(target) || args.iter().any(expr_contains_topic)
-        }
-        Expr::Index { target, index, .. } => {
-            expr_contains_topic(target) || expr_contains_topic(index)
-        }
-        Expr::InfixFunc { left, right, .. } => {
-            expr_contains_topic(left) || right.iter().any(expr_contains_topic)
-        }
-        Expr::MetaOp { left, right, .. } => expr_contains_topic(left) || expr_contains_topic(right),
-        _ => false,
-    }
-}
-
-/// Replace Whatever and nested single-arg WhateverCode with $_ (for single-arg wrapping).
-fn replace_whatever_single(expr: &Expr) -> Expr {
-    match expr {
-        e if is_whatever(e) => Expr::Var("_".to_string()),
-        // A nested single-arg WhateverCode: unwrap and reuse body (already uses $_)
-        Expr::Lambda { param, body, .. } if param == "_" => {
-            if let Some(Stmt::Expr(e)) = body.first() {
-                e.clone()
-            } else {
-                Expr::Var("_".to_string())
-            }
-        }
-        // A nested single-param WhateverCode built as AnonSubParams (its body
-        // referenced $_, so it uses a numbered param): rename that param to $_
-        // so it shares the composed closure's single topic.
-        Expr::AnonSubParams {
-            params,
-            body,
-            is_whatever_code: true,
-            ..
-        } if params.len() == 1 => {
-            if let Some(Stmt::Expr(e)) = body.first() {
-                rename_var(e, &params[0], "_")
-            } else {
-                Expr::Var("_".to_string())
-            }
-        }
-        // SmartMatch: only replace Whatever on LHS; RHS is handled at runtime.
-        Expr::Binary {
-            left,
-            op: op @ TokenKind::SmartMatch,
-            right,
-        } => Expr::Binary {
-            left: Box::new(replace_whatever_single(left)),
-            op: op.clone(),
-            right: right.clone(),
-        },
-        Expr::Binary { left, op, right } => Expr::Binary {
-            left: Box::new(replace_whatever_single(left)),
-            op: op.clone(),
-            right: Box::new(replace_whatever_single(right)),
-        },
-        Expr::Unary { op, expr } => Expr::Unary {
-            op: op.clone(),
-            expr: Box::new(replace_whatever_single(expr)),
-        },
-        Expr::PostfixOp { op, expr } => Expr::PostfixOp {
-            op: op.clone(),
-            expr: Box::new(replace_whatever_single(expr)),
-        },
-        Expr::MethodCall {
-            target,
-            name,
-            args,
-            modifier,
-            quoted,
-        } => Expr::MethodCall {
-            target: Box::new(replace_whatever_single(target)),
-            name: *name,
-            args: args.clone(),
-            modifier: *modifier,
-            quoted: *quoted,
-        },
-        Expr::DynamicMethodCall {
-            target,
-            name_expr,
-            args,
-            modifier,
-        } => Expr::DynamicMethodCall {
-            target: Box::new(replace_whatever_single(target)),
-            name_expr: name_expr.clone(),
-            args: args.clone(),
-            modifier: *modifier,
-        },
-        Expr::CallOn { target, args } => Expr::CallOn {
-            target: Box::new(replace_whatever_single(target)),
-            args: args.iter().map(replace_whatever_single).collect(),
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-            ..
-        } => Expr::Index {
-            target: Box::new(replace_whatever_single(target)),
-            index: index.clone(),
-            is_positional: *is_positional,
-        },
-        Expr::InfixFunc {
-            name,
-            left,
-            right,
-            modifier,
-        } => Expr::InfixFunc {
-            name: name.clone(),
-            left: Box::new(replace_whatever_single(left)),
-            right: right.iter().map(replace_whatever_single).collect(),
-            modifier: modifier.clone(),
-        },
-        Expr::MetaOp {
-            meta,
-            op,
-            left,
-            right,
-        } => Expr::MetaOp {
-            meta: meta.clone(),
-            op: op.clone(),
-            left: Box::new(replace_whatever_single(left)),
-            right: Box::new(replace_whatever_single(right)),
-        },
-        _ => expr.clone(),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::operators::*;
-    use super::*;
-
-    #[test]
-    fn parse_binary_add() {
-        let (rest, expr) = expression("1 + 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::Plus,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_comparison() {
-        let (rest, expr) = expression("$x == $y").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::EqEq,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_not_smartmatch() {
-        let (rest, expr) = expression("\"abc\" !~~ \"xyz\"").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::BangTilde,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_smartmatch_with_p5_subst_rhs() {
-        let (rest, expr) = expression("$path ~~ s:Perl5:g{/+} = '/'").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::SmartMatch,
-                right,
-                ..
-            } if matches!(
-                &*right,
-                Expr::Subst {
-                    pattern,
-                    replacement,
-                    global: true,
-                    perl5: true,
-                    ..
-                } if pattern == "/+" && replacement == "/"
-            )
-        ));
-    }
-
-    #[test]
-    fn parse_smartmatch_with_p5_subst_rhs_and_qq_replacement() {
-        let (rest, expr) = expression("$bar ~~ s:P5:g{$rule3}=qq{$subst}").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::SmartMatch,
-                right,
-                ..
-            } if matches!(
-                &*right,
-                Expr::AssignExpr { name, .. } if name == "_"
-            )
-        ));
-    }
-
-    #[test]
-    fn parse_smartmatch_with_regex_junction_rhs() {
-        let (rest, expr) = expression("'testing' ~~ /t/ & /s/ & /g/").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::SmartMatch,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_method_call() {
-        let (rest, expr) = expression("$x.defined").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::MethodCall { .. }));
-    }
-
-    #[test]
-    fn parse_dynamic_quoted_method_call() {
-        let (rest, expr) = expression("$_.\"$k\"()").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::DynamicMethodCall { .. }));
-    }
-
-    #[test]
-    fn parse_dynamic_sigiled_method_call() {
-        let (rest, expr) = expression("$o.$meth").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::DynamicMethodCall { .. }));
-    }
-
-    #[test]
-    fn parse_dynamic_sigiled_method_call_with_args() {
-        let (rest, expr) = expression("$o.$meth(1, 2)").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::DynamicMethodCall { args, .. } => assert_eq!(args.len(), 2),
-            _ => panic!("expected dynamic method call"),
-        }
-    }
-
-    #[test]
-    fn parse_ternary() {
-        let (rest, expr) = expression("$x ?? 1 !! 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Ternary { .. }));
-    }
-
-    #[test]
-    fn parse_ternary_allows_parenthesized_assignment_branches() {
-        let (rest, expr) = expression("$x ?? ($y = 1) !! ($y = 2)").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Ternary {
-                then_expr,
-                else_expr,
-                ..
-            } => {
-                assert!(matches!(*then_expr, Expr::Grouped(_)));
-                assert!(matches!(*else_expr, Expr::Grouped(_)));
-            }
-            _ => panic!("expected ternary expression"),
-        }
-    }
-
-    #[test]
-    fn parse_parenthesized_compound_assign_as_lvalue_expression() {
-        let (rest, expr) = expression("($x += 2) *= 3").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::AssignExpr { name, expr, .. } => {
-                assert_eq!(name, "x");
-                assert!(matches!(*expr, Expr::Binary { .. }));
-            }
-            other => panic!("expected nested AssignExpr, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn parse_parenthesized_metaassign_chain_as_lvalue_expression() {
-        let (rest, expr) = expression("(@a ||= 42) += 10").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::AssignExpr { name, expr, .. } => {
-                assert_eq!(name, "@a");
-                assert!(matches!(*expr, Expr::Binary { .. }));
-            }
-            other => panic!("expected nested AssignExpr, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn parse_junctive_op_all() {
-        assert_eq!(parse_junctive_op("?|"), Some((JunctiveOp::Or, 2)));
-        assert_eq!(parse_junctive_op("?&"), Some((JunctiveOp::And, 2)));
-        assert_eq!(parse_junctive_op("?^"), Some((JunctiveOp::Xor, 2)));
-        assert_eq!(parse_junctive_op("?"), None);
-        assert_eq!(parse_junctive_op("??"), None);
-    }
-
-    #[test]
-    fn parse_or_or_op_all() {
-        assert_eq!(parse_or_or_op("||"), Some((LogicalOp::OrOr, 2)));
-        assert_eq!(parse_or_or_op("//"), Some((LogicalOp::DefinedOr, 2)));
-        assert_eq!(parse_or_or_op("///"), None);
-        assert_eq!(parse_or_or_op("|"), None);
-    }
-
-    #[test]
-    fn parse_and_and_op_all() {
-        assert_eq!(parse_and_and_op("&&"), Some((LogicalOp::AndAnd, 2)));
-        assert_eq!(parse_and_and_op("&"), None);
-    }
-
-    #[test]
-    fn parse_word_logical_op_all() {
-        assert_eq!(parse_word_logical_op("or "), Some((LogicalOp::Or, 2)));
-        assert_eq!(parse_word_logical_op("and "), Some((LogicalOp::And, 3)));
-        assert_eq!(parse_word_logical_op("or_foo"), None);
-        assert_eq!(parse_word_logical_op("and_bar"), None);
-        assert_eq!(parse_word_logical_op("oracle"), None);
-    }
-
-    #[test]
-    fn parse_logical_operators_in_expr() {
-        // Test word forms
-        let (rest, expr) = expression("1 or 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::OrWord,
-                ..
-            }
-        ));
-
-        let (rest, expr) = expression("1 and 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::AndWord,
-                ..
-            }
-        ));
-
-        // Test symbolic forms
-        let (rest, expr) = expression("1 || 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::OrOr,
-                ..
-            }
-        ));
-
-        let (rest, expr) = expression("1 && 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::AndAnd,
-                ..
-            }
-        ));
-
-        let (rest, expr) = expression("1 // 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::SlashSlash,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_junctive_operators_in_expr() {
-        let inputs = [("1 ?| 2", "?|"), ("1 ?& 2", "?&"), ("1 ?^ 2", "?^")];
-
-        for (input, expected_op) in inputs {
-            let (rest, expr) = expression(input).unwrap();
-            assert_eq!(rest, "");
-            if let Expr::Binary { op, .. } = expr {
-                match op {
-                    TokenKind::Ident(s) => assert_eq!(s, expected_op),
-                    _ => panic!("Expected Ident token for {}", expected_op),
-                }
-            } else {
-                panic!("Expected Binary expression");
-            }
-        }
-    }
-
-    #[test]
-    fn parse_one_junction_operator_without_spaces() {
-        let (rest, expr) = expression("3^2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::Caret,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_fat_arrow_chains_right_associatively() {
-        let (rest, expr) = expression("1 => 2 => 3 => 4").unwrap();
-        assert_eq!(rest, "");
-        // Non-bareword key => outermost wrapped in PositionalPair
-        // Inner chain pairs are raw Binary (not wrapped)
-        match expr {
-            Expr::PositionalPair(inner) => match *inner {
-                Expr::Binary {
-                    op: TokenKind::FatArrow,
-                    left,
-                    right,
-                } => {
-                    assert!(matches!(*left, Expr::Literal(Value::Int(1))));
-                    match *right {
-                        Expr::Binary {
-                            op: TokenKind::FatArrow,
-                            left: right_left,
-                            right: right_right,
-                        } => {
-                            assert!(matches!(*right_left, Expr::Literal(Value::Int(2))));
-                            match *right_right {
-                                Expr::Binary {
-                                    op: TokenKind::FatArrow,
-                                    left: tail_left,
-                                    right: tail_right,
-                                } => {
-                                    assert!(matches!(*tail_left, Expr::Literal(Value::Int(3))));
-                                    assert!(matches!(*tail_right, Expr::Literal(Value::Int(4))));
-                                }
-                                _ => panic!("expected final fat-arrow pair"),
-                            }
-                        }
-                        _ => panic!("expected nested fat-arrow pair"),
-                    }
-                }
-                _ => panic!("expected fat-arrow pair inside PositionalPair"),
-            },
-            _ => panic!("expected PositionalPair for non-bareword key"),
-        }
-    }
-
-    #[test]
-    fn parse_fat_arrow_chain_in_call_arguments() {
-        let (rest, expr) = expression("is($list, 1 => 2 => 3 => 4, \"x\")").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "is");
-                assert!(args.len() >= 3);
-                // Non-bareword key => outermost PositionalPair wrapping
-                // Inner chain pairs are raw Binary
-                match &args[1] {
-                    Expr::PositionalPair(inner) => match inner.as_ref() {
-                        Expr::Binary {
-                            op: TokenKind::FatArrow,
-                            right,
-                            ..
-                        } => {
-                            assert!(matches!(
-                                right.as_ref(),
-                                Expr::Binary {
-                                    op: TokenKind::FatArrow,
-                                    ..
-                                }
-                            ));
-                        }
-                        _ => panic!("expected chained fat-arrow in second argument"),
-                    },
-                    _ => panic!("expected PositionalPair for non-bareword key"),
-                }
-            }
-            _ => panic!("expected call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_upto_with_infinity_literal() {
-        let (rest, expr) = expression("^∞").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Binary { left, op, right } => {
-                assert!(matches!(*left, Expr::Literal(Value::Int(0))));
-                assert!(matches!(op, TokenKind::DotDotCaret));
-                assert!(matches!(
-                    *right,
-                    Expr::Literal(Value::Int(_))
-                        | Expr::Literal(Value::BigInt(_))
-                        | Expr::Literal(Value::Num(_))
-                ));
-            }
-            _ => panic!("expected upto range expression"),
-        }
-    }
-
-    #[test]
-    fn parse_upto_with_negative_literal() {
-        let (rest, expr) = expression("^-1").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Binary { left, op, right } => {
-                assert!(matches!(*left, Expr::Literal(Value::Int(0))));
-                assert!(matches!(op, TokenKind::DotDotCaret));
-                assert!(matches!(
-                    *right,
-                    Expr::Unary {
-                        op: TokenKind::Minus,
-                        ..
-                    }
-                ));
-            }
-            _ => panic!("expected upto range expression"),
-        }
-    }
-
-    #[test]
-    fn parse_topical_dot_angle_expression() {
-        let (rest, expr) = expression(".<a>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Index { .. }));
-    }
-
-    #[test]
-    fn parse_topical_dot_brace_expression() {
-        let (rest, expr) = expression(".{'a'}").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Index { .. }));
-    }
-
-    #[test]
-    fn parse_dynamic_pseudostash_dot_angle_with_dynamic_var() {
-        let (rest, expr) = expression("DYNAMIC::.<$*x80>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Index { target, index, .. }
-                if matches!(target.as_ref(), Expr::PseudoStash(s) if s.as_str() == "DYNAMIC::")
-                && matches!(index.as_ref(), Expr::Literal(Value::Str(s)) if s.as_str() == "$*x80")
-        ));
-    }
-
-    #[test]
-    fn expression_memo_reuses_result() {
-        reset_expression_memo();
-        let (rest, expr) = expression("1 + 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Binary { .. }));
-        let (rest2, expr2) = expression("1 + 2").unwrap();
-        assert_eq!(rest2, "");
-        assert!(matches!(expr2, Expr::Binary { .. }));
-    }
-
-    #[test]
-    fn comparison_requires_rhs_expression() {
-        let err = expression("1 <").unwrap_err();
-        assert!(
-            err.message()
-                .contains("expression after comparison operator")
-        );
-    }
-
-    #[test]
-    fn parse_strict_not_equal_operator() {
-        let (rest, expr) = expression("1 !=== 2").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Unary {
-                op: TokenKind::Bang,
-                expr,
-            } => match *expr {
-                Expr::Binary {
-                    op: TokenKind::EqEqEq,
-                    ..
-                } => {}
-                _ => panic!("Expected !=== to lower to !(===)"),
-            },
-            _ => panic!("Expected Binary expression"),
-        }
-    }
-
-    #[test]
-    fn parse_negated_comparison_meta_operators() {
-        let (rest, expr) = expression("2 !== 3").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Unary {
-                op: TokenKind::Bang,
-                expr,
-            } => match *expr {
-                Expr::Binary {
-                    op: TokenKind::EqEq,
-                    ..
-                } => {}
-                _ => panic!("Expected !== to lower to !(==)"),
-            },
-            _ => panic!("Expected Unary expression"),
-        }
-
-        let (rest, expr) = expression("$a !eq $b").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Unary {
-                op: TokenKind::Bang,
-                expr,
-            } => match *expr {
-                Expr::Binary {
-                    op: TokenKind::Ident(op),
-                    ..
-                } => assert_eq!(op, "eq"),
-                _ => panic!("Expected !eq to lower to !(eq)"),
-            },
-            _ => panic!("Expected Unary expression"),
-        }
-    }
-
-    #[test]
-    fn parse_container_not_equal_operator() {
-        let (rest, expr) = expression("$a !=:= $b").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Unary {
-                op: TokenKind::Bang,
-                expr,
-            } => match *expr {
-                Expr::Binary {
-                    op: TokenKind::Ident(op),
-                    ..
-                } => assert_eq!(op, "=:="),
-                _ => panic!("Expected !=:= to lower to !(=:=)"),
-            },
-            _ => panic!("Expected unary ! expression"),
-        }
-    }
-
-    #[test]
-    fn parse_cross_with_container_not_equal_operator() {
-        let (rest, expr) = expression("$a X!=:= $b").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MetaOp { meta, op, .. } => {
-                assert_eq!(meta, "X");
-                assert_eq!(op, "!=:=");
-            }
-            _ => panic!("Expected cross meta operator expression"),
-        }
-    }
-
-    #[test]
-    fn parse_cross_dot_string_reports_obsolete_error() {
-        let err = expression("3 X. \"foo\"").unwrap_err();
-        assert!(err.message().contains("X::Obsolete"));
-    }
-
-    #[test]
-    fn chained_comparison_requires_rhs_expression() {
-        let err = expression("1 < 2 <").unwrap_err();
-        assert!(err.message().contains("chained comparison operator"));
-    }
-
-    #[test]
-    fn range_requires_rhs_expression() {
-        let err = expression("1 ..").unwrap_err();
-        assert!(err.message().contains("range RHS"));
-    }
-
-    #[test]
-    fn ternary_requires_then_and_else_expression() {
-        let err_then = expression("$x ??").unwrap_err();
-        assert!(err_then.message().contains("then-expression"));
-
-        let err_else = expression("$x ?? 1 !!").unwrap_err();
-        assert!(err_else.message().contains("else-expression"));
-    }
-
-    #[test]
-    fn additive_requires_rhs_expression() {
-        let err = expression("1 +").unwrap_err();
-        assert!(err.message().contains("additive operator"));
-    }
-
-    #[test]
-    fn parse_postfix_method_requires_name() {
-        let err = expression("$x.").unwrap_err();
-        assert!(err.message().contains("method name"));
-    }
-
-    #[test]
-    fn parse_postfix_colon_args_require_expression() {
-        let err = expression("$x.foo:").unwrap_err();
-        assert!(err.message().contains("expected"));
-    }
-
-    #[test]
-    fn parse_private_postfix_colon_args() {
-        let (rest, expr) = expression("self.bless!SET-SELF: pulled, i").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MethodCall {
-                target,
-                name,
-                args,
-                modifier,
-                ..
-            } => {
-                assert_eq!(name, "SET-SELF");
-                assert_eq!(modifier, Some('!'));
-                assert_eq!(args.len(), 2);
-                assert!(matches!(args[0], Expr::BareWord(ref n) if n.as_str() == "pulled"));
-                assert!(matches!(args[1], Expr::BareWord(ref n) if n.as_str() == "i"));
-                assert!(matches!(
-                    *target,
-                    Expr::MethodCall {
-                        name,
-                        modifier: None,
-                        ..
-                    } if name == "bless"
-                ));
-            }
-            _ => panic!("expected private method call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_postfix_angle_index_requires_closing() {
-        let err = expression("$x<foo").unwrap_err();
-        assert!(err.message().contains("closing '>'"));
-    }
-
-    #[test]
-    fn parse_postfix_angle_index_zen() {
-        let (rest, expr) = expression("$x<>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::MethodCall {
-                target,
-                name,
-                args,
-                ..
-            } if name == "__mutsu_zen_angle"
-                && args.is_empty()
-                && matches!(*target, Expr::Var(ref n) if n.as_str() == "x")
-        ));
-    }
-
-    #[test]
-    fn parse_postfix_angle_index_zen_values_adverb() {
-        let (rest, expr) = expression("%h<>:v").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MethodCall {
-                target, name, args, ..
-            } => {
-                assert_eq!(name, "values");
-                assert!(args.is_empty());
-                assert!(matches!(*target, Expr::HashVar(ref n) if n.as_str() == "h"));
-            }
-            _ => panic!("expected method call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_postfix_angle_index_multiple_keys() {
-        let (rest, expr) = expression("%h<a b c>").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Index { index, .. } => {
-                assert!(matches!(*index, Expr::ArrayLiteral(ref items) if items.len() == 3));
-            }
-            _ => panic!("expected index expression"),
-        }
-    }
-
-    #[test]
-    fn parse_postfix_call_adverb_block_as_arg() {
-        let (rest, expr) = expression("foo():{ 42 }").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "foo");
-                assert_eq!(args.len(), 2);
-                assert!(args.iter().any(|arg| matches!(arg, Expr::AnonSub { .. })));
-                assert!(args.iter().any(|arg| matches!(
-                    arg,
-                    Expr::Binary { left, op: crate::token_kind::TokenKind::FatArrow, .. }
-                    if matches!(left.as_ref(), Expr::Literal(crate::value::Value::Str(s)) if s.as_str() == "__mutsu_test_callsite_line")
-                )));
-            }
-            _ => panic!("expected call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_postfix_call_adverb_colonpairs_as_args() {
-        let (rest, expr) = expression("fiddle():x(\"a\"):y").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "fiddle");
-                assert_eq!(args.len(), 3);
-                assert!(
-                    args.iter()
-                        .filter(|arg| matches!(arg, Expr::Binary { .. }))
-                        .count()
-                        >= 3
-                );
-                assert!(args.iter().any(|arg| matches!(
-                    arg,
-                    Expr::Binary { left, op: crate::token_kind::TokenKind::FatArrow, .. }
-                    if matches!(left.as_ref(), Expr::Literal(crate::value::Value::Str(s)) if s.as_str() == "__mutsu_test_callsite_line")
-                )));
-            }
-            _ => panic!("expected call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_subscript_adverb_does_not_become_call_adverb() {
-        let (rest, expr) = expression("say @a[0]:p").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "say");
-                assert_eq!(args.len(), 1);
-                assert!(
-                    matches!(args[0], Expr::Call { ref name, .. } if name == "__mutsu_subscript_adverb")
-                );
-            }
-            _ => panic!("expected call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_hash_subscript_adverb_does_not_become_call_adverb() {
-        let (rest, expr) = expression("say %h<a>:kv").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "say");
-                assert_eq!(args.len(), 1);
-                assert!(
-                    matches!(args[0], Expr::Call { ref name, .. } if name == "__mutsu_subscript_adverb")
-                );
-            }
-            _ => panic!("expected call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_listop_block_then_colon_args() {
-        // In Raku, `map { block }: @list` is invocant-colon syntax:
-        // it means `{ block }.map(@list)`, i.e., a method call on the block.
-        let (rest, expr) = expression("map { $_ * 2 }: @list").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MethodCall { name, args, .. } => {
-                assert_eq!(name, "map");
-                assert_eq!(args.len(), 1);
-                assert!(matches!(args[0], Expr::ArrayVar(ref n) if n.as_str() == "list"));
-            }
-            _ => panic!("expected method call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_method_colon_arg_pointy_with_return_type() {
-        let (rest, expr) = expression("@a.map: -> \\x --> Int { x }").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MethodCall { name, args, .. } => {
-                assert_eq!(name, "map");
-                assert_eq!(args.len(), 1);
-                assert!(matches!(args[0], Expr::Lambda { ref param, .. } if param.as_str() == "x"));
-            }
-            _ => panic!("expected method call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_method_colon_args_adjacent_colonpairs() {
-        let src = "IO::Path.new: :volume<foo> :dirname<bar> :basename<ber> :SPEC(IO::Spec::Win32)";
-        let (rest, expr) = expression(src).unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MethodCall { name, args, .. } => {
-                assert_eq!(name, "new");
-                assert_eq!(args.len(), 4);
-            }
-            _ => panic!("expected method call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_indir_with_hyphenated_call_arg_and_block() {
-        let (rest, expr) = expression("indir make-temp-dir, { 42 }").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "indir");
-                assert_eq!(args.len(), 2);
-            }
-            _ => panic!("expected call expression"),
-        }
-    }
-
-    #[test]
-    fn parse_paren_expr_with_space_dot_method_chain() {
-        let (rest, expr) = expression("(^3 .map: -> \\x --> Int { $i++ }).sink").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MethodCall { name, .. } => assert_eq!(name, "sink"),
-            _ => panic!("expected outer method call"),
-        }
-    }
-
-    #[test]
-    fn parse_hash_literal_with_hash_spread_expr() {
-        let (rest, expr) = expression("{year => 1984, %args}").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "hash");
-                assert_eq!(args.len(), 2);
-            }
-            _ => panic!("expected lowered hash() call"),
-        }
-    }
-
-    #[test]
-    fn parse_slip_prefix_with_parenthesized_expr() {
-        let (rest, expr) = expression("|(f)").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_slip_prefix_with_reduction_expr() {
-        let (rest, expr) = expression("|[\\+] 1..*").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_slip_prefix_with_topic_method_call() {
-        let (rest, expr) = expression("|.value").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                expr
-            } if matches!(*expr, Expr::MethodCall { .. })
-        ));
-    }
-
-    #[test]
-    fn parse_slip_prefix_with_upto_operator() {
-        let (rest, expr) = expression("|^2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_parenthesized_slip_prefix_with_method_chain() {
-        let (rest, expr) = expression("(|^2).Seq").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::MethodCall { name, .. } if name == "Seq"
-        ));
-    }
-
-    #[test]
-    fn parse_take_listop_with_xx_expression_argument() {
-        let (rest, expr) = expression("take (@b.shift xx 2) xx 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Call { name, .. } if name == "take"));
-    }
-
-    #[test]
-    fn parse_slip_prefix_with_quote_word_list() {
-        let (rest, expr) = expression("|<a b>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_hyper_prefix_slip_left_on_angle_list() {
-        let (rest, expr) = expression("|<< <a x y z>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_hyper_prefix_slip_right_on_angle_list() {
-        let (rest, expr) = expression("|>> <a x y z>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_slip_prefix_with_space_before_french_quote_list() {
-        let (rest, expr) = expression("| «echo test»").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                expr
-            } if matches!(
-                *expr,
-                Expr::ArrayLiteral(ref items) if items.len() == 2
-            ) || matches!(
-                *expr,
-                Expr::Call { ref name, ref args }
-                    if name.resolve() == "list" && args.len() == 2
-            )
-        ));
-    }
-
-    #[test]
-    fn parse_slip_prefix_with_space_before_var() {
-        let (rest, expr) = expression("|   @cmd").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Pipe,
-                expr
-            } if matches!(*expr, Expr::ArrayVar(ref name) if name == "cmd")
-        ));
-    }
-
-    #[test]
-    fn parse_prefix_boolify_codevar_method_call() {
-        let (rest, expr) = expression("?&foo.cando($c)").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Unary {
-                op: TokenKind::Question,
-                expr,
-            } => match *expr {
-                Expr::MethodCall { target, .. } => {
-                    assert!(matches!(*target, Expr::CodeVar(ref name) if name == "foo"));
-                }
-                _ => panic!("expected method call operand"),
-            },
-            _ => panic!("expected prefix boolify expression"),
-        }
-    }
-
-    #[test]
-    fn parse_hyper_prefix_metaop_negate() {
-        let (rest, expr) = expression("-« ([1, 2], 3)").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "__mutsu_hyper_prefix");
-                assert_eq!(args.len(), 2);
-                assert!(matches!(args[0], Expr::Literal(Value::Str(ref s)) if s.as_str() == "-"));
-            }
-            _ => panic!("expected hyper prefix metaop call"),
-        }
-    }
-
-    #[test]
-    fn parse_parenthesized_sequence_with_following_smartmatch() {
-        let (rest, expr) = expression("(\"a\"...* ~~ / z /)").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Binary { .. }));
-    }
-
-    #[test]
-    fn parse_parenthesized_sequence_with_empty_paren_seed_and_postfix_index() {
-        let (rest, expr) = expression("(() ... *)[0]").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Index { .. }));
-    }
-
-    #[test]
-    fn parse_s_metaop_infix_and() {
-        let (rest, expr) = expression("1 S& 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Binary { .. }));
-    }
-
-    #[test]
-    fn parse_zip_meta_with_parenthesized_set_intersection() {
-        let (rest, expr) = expression("1..3 Z(&) 2..4").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MetaOp { meta, op, .. } => {
-                assert_eq!(meta, "Z");
-                assert_eq!(op, "(&)");
-            }
-            _ => panic!("expected zip meta op"),
-        }
-    }
-
-    #[test]
-    fn parse_zip_meta_with_unicode_set_intersection() {
-        let (rest, expr) = expression("1..3 Z∩ 2..4").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::MetaOp { meta, op, .. } => {
-                assert_eq!(meta, "Z");
-                assert_eq!(op, "∩");
-            }
-            _ => panic!("expected zip meta op"),
-        }
-    }
-
-    #[test]
-    fn parse_unicode_set_union_infix() {
-        let (rest, expr) = expression("1 ∪ 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::SetUnion,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_whatever_with_unicode_set_union_infix() {
-        let (rest, expr) = expression("* ∪ *").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::AnonSubParams { body, .. }
-                if matches!(
-                    body.as_slice(),
-                    [Stmt::Expr(Expr::Binary { op: TokenKind::SetUnion, .. })]
-                )
-        ));
-    }
-
-    #[test]
-    fn parse_zip_metaop_with_set_union_ascii() {
-        let (rest, expr) = expression("1..3 Z(|) 2..4").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::MetaOp { ref meta, ref op, .. } if meta == "Z" && op == "(|)"
-        ));
-    }
-
-    #[test]
-    fn parse_zip_metaop_with_set_union_unicode() {
-        let (rest, expr) = expression("1..3 Z∪ 2..4").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::MetaOp { ref meta, ref op, .. } if meta == "Z" && op == "∪"
-        ));
-    }
-
-    #[test]
-    fn parse_cross_metaop_with_bitshift_right() {
-        let (rest, expr) = expression("1 X+> 2").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::MetaOp { ref meta, ref op, .. } if meta == "X" && op == "+>"
-        ));
-    }
-
-    #[test]
-    fn parse_grouped_sequences_as_metaop_operands() {
-        let (rest, expr) = expression("(1 ... *) Z~ ('a' ... 'z')").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::MetaOp { ref meta, ref op, .. } if meta == "Z" && op == "~"
-        ));
-    }
-
-    #[test]
-    fn parse_bracket_infix_func_chain_with_same_operator() {
-        let (rest, expr) = expression("2 [&foo] 3 [&foo] 4").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::InfixFunc { ref name, modifier: None, .. } if name == "foo"
-        ));
-    }
-
-    #[test]
-    fn parse_reverse_metaop_on_comma_list_operator() {
-        let (rest, expr) = expression("1 R, 2 R, 3 R, 4").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::MetaOp { .. }));
-    }
-
-    #[test]
-    fn parse_reverse_metaop_andand() {
-        let (rest, expr) = expression("$x R&& 0").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::MetaOp { .. }));
-    }
-
-    #[test]
-    fn parse_reverse_metaop_oror() {
-        let (rest, expr) = expression("$x R|| 1").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::MetaOp { .. }));
-    }
-
-    #[test]
-    fn parse_reverse_meta_compound_assign_rhs_target() {
-        let (rest, expr) = expression("10 R+= ($a ||= 42)").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::DoBlock { .. } | Expr::AssignExpr { .. }
-        ));
-    }
-
-    #[test]
-    fn parse_reverse_meta_compound_assign_non_lvalue_target() {
-        let (rest, expr) = expression("($a ||= 42) R+= 10").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::DoBlock { .. }));
-    }
-
-    #[test]
-    fn parse_reverse_dot_concat_reports_obsolete() {
-        let err = expression("3 R. \"foo\"").unwrap_err();
-        assert!(err.message().contains("X::Obsolete"));
-    }
-
-    #[test]
-    fn parse_pair_lvalue_colonparen_form() {
-        let (rest, expr) = expression(":(:$a is raw)").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Literal(Value::Instance { ref class_name, .. }) if class_name == "Signature"
-        ));
-    }
-
-    #[test]
-    fn parse_custom_postfix_operator_call() {
-        let (rest, expr) = expression("$base!").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "postfix:<!>");
-                assert_eq!(args.len(), 1);
-                assert!(matches!(args[0], Expr::Var(ref n) if n.as_str() == "base"));
-            }
-            _ => panic!("expected postfix operator call"),
-        }
-    }
-
-    #[test]
-    fn parse_unicode_custom_postfix_operator_call() {
-        let (rest, expr) = expression("3§").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "postfix:<§>");
-                assert_eq!(args.len(), 1);
-                assert!(matches!(args[0], Expr::Literal(Value::Int(3))));
-            }
-            _ => panic!("expected unicode postfix operator call"),
-        }
-    }
-
-    #[test]
-    fn parse_dot_custom_postfix_operator_call() {
-        let (rest, expr) = expression("5.!").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "postfix:<!>");
-                assert_eq!(args.len(), 1);
-                assert!(matches!(args[0], Expr::Literal(Value::Int(5))));
-            }
-            _ => panic!("expected dot-postfix operator call"),
-        }
-    }
-
-    #[test]
-    fn parse_dot_postfix_increment() {
-        let (rest, expr) = expression("$x.++").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::PostfixOp {
-                op: TokenKind::PlusPlus,
-                expr,
-            } => {
-                assert!(matches!(*expr, Expr::Var(ref n) if n.as_str() == "x"));
-            }
-            _ => panic!("expected dot-postfix increment"),
-        }
-    }
-
-    #[test]
-    fn parse_dot_hyper_postfix_increment() {
-        let (rest, expr) = expression("$x.>>++").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::HyperMethodCall {
-                target, name, args, ..
-            } => {
-                assert!(matches!(*target, Expr::Var(ref n) if n.as_str() == "x"));
-                assert_eq!(name, "postfix:<++>");
-                assert!(args.is_empty());
-            }
-            _ => panic!("expected hyper postfix increment"),
-        }
-    }
-
-    #[test]
-    fn parse_hyper_index_without_dot() {
-        let (rest, expr) = expression("%h{*}»[1]").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::HyperMethodCall {
-                target, name, args, ..
-            } => {
-                assert_eq!(name, "AT-POS");
-                assert_eq!(args.len(), 1);
-                assert!(matches!(*target, Expr::Index { .. }));
-            }
-            _ => panic!("expected hyper index call"),
-        }
-    }
-
-    #[test]
-    fn parse_hyper_index_with_dot() {
-        let (rest, expr) = expression("%h{*}».[1]").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::HyperMethodCall {
-                target, name, args, ..
-            } => {
-                assert_eq!(name, "AT-POS");
-                assert_eq!(args.len(), 1);
-                assert!(matches!(*target, Expr::Index { .. }));
-            }
-            _ => panic!("expected hyper dot-index call"),
-        }
-    }
-
-    #[test]
-    fn parse_dot_ampersand_block_call() {
-        let (rest, expr) = expression("$m.&{ 3 }").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::DynamicMethodCall {
-                target,
-                name_expr,
-                args,
-                ..
-            } => {
-                assert!(matches!(*target, Expr::Var(ref n) if n.as_str() == "m"));
-                assert!(matches!(*name_expr, Expr::AnonSub { .. }));
-                assert!(args.is_empty());
-            }
-            _ => panic!("expected dynamic method call"),
-        }
-    }
-
-    #[test]
-    fn parse_hyper_dot_ampersand_block_call() {
-        let (rest, expr) = expression("$m>>.&{ 3 }").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::HyperMethodCallDynamic {
-                target,
-                name_expr,
-                args,
-                ..
-            } => {
-                assert!(matches!(*target, Expr::Var(ref n) if n.as_str() == "m"));
-                assert!(matches!(*name_expr, Expr::AnonSub { .. }));
-                assert!(args.is_empty());
-            }
-            _ => panic!("expected hyper dynamic method call"),
-        }
-    }
-
-    #[test]
-    fn parse_hyper_method_with_unspace_after_operator() {
-        let (rest, expr) = expression("$m»\\\n.foo").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::HyperMethodCall { target, name, .. } => {
-                assert!(matches!(*target, Expr::Var(ref n) if n.as_str() == "m"));
-                assert_eq!(name, "foo");
-            }
-            _ => panic!("expected hyper method call"),
-        }
-    }
-
-    #[test]
-    fn parse_array_slice_assignment_with_comma_rhs() {
-        let (rest, expr) = expression("@a[0,1] = 10,20").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::IndexAssign { value, .. } => match *value {
-                Expr::ArrayLiteral(items) => {
-                    assert_eq!(items.len(), 2);
-                    assert!(matches!(items[0], Expr::Literal(Value::Int(10))));
-                    assert!(matches!(items[1], Expr::Literal(Value::Int(20))));
-                }
-                other => panic!("expected ArrayLiteral RHS, got {other:?}"),
-            },
-            _ => panic!("expected IndexAssign expression"),
-        }
-    }
-
-    #[test]
-    fn parse_hash_slice_assignment_with_comma_rhs() {
-        let (rest, expr) = expression("%h{1,2} = \"one\", \"two\"").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::IndexAssign { value, .. } => match *value {
-                Expr::ArrayLiteral(items) => {
-                    assert_eq!(items.len(), 2);
-                    assert!(
-                        matches!(items[0], Expr::Literal(Value::Str(ref s)) if s.as_str() == "one")
-                    );
-                    assert!(
-                        matches!(items[1], Expr::Literal(Value::Str(ref s)) if s.as_str() == "two")
-                    );
-                }
-                other => panic!("expected ArrayLiteral RHS, got {other:?}"),
-            },
-            _ => panic!("expected IndexAssign expression"),
-        }
-    }
-
-    #[test]
-    fn parse_dot_postfix_decrement() {
-        let (rest, expr) = expression("$x.--").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::PostfixOp {
-                op: TokenKind::MinusMinus,
-                expr,
-            } => {
-                assert!(matches!(*expr, Expr::Var(ref n) if n.as_str() == "x"));
-            }
-            _ => panic!("expected dot-postfix decrement"),
-        }
-    }
-
-    #[test]
-    fn parse_dot_method_name_then_postfix_decrement() {
-        let (rest, expr) = expression("$obj.b--").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::PostfixOp {
-                op: TokenKind::MinusMinus,
-                expr,
-            } => match *expr {
-                Expr::MethodCall { name, .. } => assert_eq!(name.resolve(), "b"),
-                _ => panic!("expected MethodCall inside postfix decrement"),
-            },
-            _ => panic!("expected postfix decrement on method call"),
-        }
-    }
-
-    #[test]
-    fn parse_ascii_minus_on_angle_complex_literal() {
-        let (rest, expr) = expression("-<42+2i>").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Unary {
-                op: TokenKind::Minus,
-                expr,
-            } => {
-                // Single-element <42+2i> produces plain Complex (not ComplexStr allomorph)
-                assert!(matches!(*expr, Expr::Literal(Value::Complex(42.0, 2.0))));
-            }
-            _ => panic!("expected unary minus expression"),
-        }
-    }
-
-    #[test]
-    fn parse_is_deeply_with_unicode_and_ascii_minus_complex_literals() {
-        let (rest, expr) = expression("is-deeply −<42+2i>, -<42+2i>").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "is-deeply");
-                assert!(args.len() >= 2);
-                assert!(matches!(
-                    args[0],
-                    Expr::Unary {
-                        op: TokenKind::Minus,
-                        ..
-                    }
-                ));
-                assert!(matches!(
-                    args[1],
-                    Expr::Unary {
-                        op: TokenKind::Minus,
-                        ..
-                    }
-                ));
-            }
-            _ => panic!("expected is-deeply call"),
-        }
-    }
-
-    #[test]
-    fn parse_expr_listop_with_topic_method_first_arg() {
-        let (rest, expr) = expression("is-deeply .bool-only, True, 'ok'").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Call { name, args } => {
-                assert_eq!(name, "is-deeply");
-                assert!(args.len() >= 3);
-                assert!(matches!(args[0], Expr::MethodCall { .. }));
-            }
-            _ => panic!("expected is-deeply call"),
-        }
-    }
-
-    #[test]
-    fn parse_unary_plus_on_topic_method_call() {
-        let (rest, expr) = expression("+.lines").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Unary {
-                op: TokenKind::Plus,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn parse_x_with_bare_whatever_rhs_without_whatevercode_wrap() {
-        let (rest, expr) = expression("'a' x *").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::Ident(ref name),
-                ..
-            } if name == "x"
-        ));
-    }
-
-    #[test]
-    fn parse_xx_with_bare_whatever_rhs_without_whatevercode_wrap() {
-        let (rest, expr) = expression("42 xx *").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::Binary {
-                op: TokenKind::Ident(ref name),
-                ..
-            } if name == "xx"
-        ));
-    }
-
-    #[test]
-    fn parse_xx_with_bare_whatever_and_postfix_index_without_wrap() {
-        let (rest, expr) = expression("((2,4,6) xx *)[^2]").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::Index { .. }));
-    }
-
-    #[test]
-    fn parse_ternary_with_expr_listop_branches() {
-        let (rest, expr) =
-            expression(".can('bool-only') ?? is-deeply .bool-only, True, 'ok' !! skip 'x'")
-                .unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::Ternary {
-                then_expr,
-                else_expr,
-                ..
-            } => {
-                assert!(matches!(*then_expr, Expr::Call { .. }));
-                assert!(matches!(*else_expr, Expr::Call { .. }));
-            }
-            _ => panic!("expected ternary expression"),
-        }
-    }
-
-    #[test]
-    fn parse_ampersand_sigiled_callable_invocation() {
-        let (rest, expr) = expression("&$x()").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            expr,
-            Expr::CallOn { target, args } if args.is_empty() && matches!(*target, Expr::Var(ref n) if n.as_str() == "x")
-        ));
-    }
-
-    #[test]
-    fn parse_ampersand_infix_operator_reference_double_angles() {
-        let (rest, expr) = expression("&infix:<<(<=)>>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::CodeVar(ref name) if name == "infix:<(<=)>"));
-    }
-
-    #[test]
-    fn parse_ampersand_infix_operator_reference_unicode_symbol() {
-        let (rest, expr) = expression("&infix:<⊆>").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(expr, Expr::CodeVar(ref name) if name == "infix:<⊆>"));
-    }
 }
